Move metron-rest hbase client
